feature: update Bills endpoints to handle user actions

+++ src/CollectPay/CollectPay.Application/BillAggregate/Commands/Bills/RemoveBill/RemoveBillCommandHandler.cs
@@ -23,6 +23,11 @@ public class RemoveBillCommandHandler : ICommandHandler<RemoveBillCommand, Delet
 			return BillErrors.BillNotFound;
 		}
 
+		if (bill.CreatorId != request.UserId)
+		{
+			return BillErrors.CannotBeRemovedByNotOwner;
+		}
+
 		await _billRepository.RemoveAsync(bill, cancellationToken);
 		return Result.Deleted;
 	}
